Correctly escape arguments for system()
To properly handle special characters such as \ and ; in arguments
forwarded to subprocesses, the arguments have to be escaped before being
passed so system(). The correct way to do this depends on the OS.

On Unix, wrapping all arguments, including argv[0], into single quotes
is both maximally safe and compatible.

On Windows, single quotes are not sufficient to escape ;. Instead, we
use double quotes. A peculiarity of system() is that it strips the first
and last double quote if the entire command is wrapped into double
quotes, which messes up the quoting if every argument is quoted
individually. We work around this by not quoting argv[0]. Since it is
safe to do and simplifies the code, we also do not quote argv[0] on
Unix.