Do not redundantly copy fuzzer input
Jazzer fuzz targets operate on a byte[] view of the fuzzer input backed
by a copy created by the JVM, so even native access to the fuzzer input
cannot result in an (accidentally) overwritten input. Making a copy of
the input for every fuzz target execution is thus unnecessary overhead
in the context of Jazzer.